flask_ticket/requirements.txt
flask
requests
dotenv
psycopg2-binary
numpy
opencv-python
bcrypt
PyJWT
